# api/openapi.yaml
openapi: 3.0.3
info:
  title: RefMD API
  description: Real-time collaborative markdown editor API with CRDT support
  version: 1.1.0
  contact:
    name: RefMD Team
    email: support@refmd.io
servers:
  - url: http://localhost:8888/api
    description: Development server
  - url: https://api.refmd.io/api
    description: Production server
paths:
...
  /documents/{id}:
    get:
      tags:
        - Documents
      summary: Get document by ID
      operationId: getDocument
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          description: Share token for accessing shared documents
          schema:
            type: string
      responses:
        '200':
          description: Document retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Documents
      summary: Update document
      operationId: updateDocument
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          description: Share token for updating shared documents
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Document updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Documents
      summary: Delete document
      operationId: deleteDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /files:
    get:
      tags:
        - Files
      summary: List files for a document
      operationId: listFiles
      security:
        - bearerAuth: []
      parameters:
        - name: document_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/File'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /files/upload:
    post:
      tags:
        - Files
      summary: Upload file
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                document_id:
                  type: string
                  format: uuid
              required:
                - file
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /files/{id}:
    get:
      tags:
        - Files
      summary: Download file
      operationId: downloadFile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File retrieved successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Files
      summary: Delete file
      operationId: deleteFile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: File deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /files/documents/{filename}:
    get:
      tags:
        - Files
      summary: Download file by filename
      description: Download file by filename from document directory. Supports public access with share token.
      operationId: downloadFileByName
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
        - name: document_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          description: Share token for accessing files in shared documents
          schema:
            type: string
      responses:
        '200':
          description: File retrieved successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===== Sharing =====
...
  /git/commits/file/{file_path}:
    get:
      tags:
        - Git Sync
      summary: Get file commit history
      operationId: getFileCommitHistory
      security:
        - bearerAuth: []
      parameters:
        - name: file_path
          in: path
          required: true
          schema:
            type: string
          description: Path to the file relative to the repository root
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          description: Maximum number of commits to return
      responses:
        '200':
          description: File commit history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GitCommit'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /git/diff/files/{file_path}:
    get:
      tags:
        - Git Sync
      summary: Get file diff
      description: Get uncommitted changes for a specific file
      operationId: getFileDiff
      security:
        - bearerAuth: []
      parameters:
        - name: file_path
          in: path
          required: true
          schema:
            type: string
          description: Path to the file relative to the repository root
      responses:
        '200':
          description: File diff retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiffResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /git/diff/commits/{from}/{to}/file/{file_path}:
    get:
      tags:
        - Git Sync
      summary: Get file diff between commits
      description: Get diff for a specific file between two commits
      operationId: getFileCommitDiff
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: path
          required: true
          schema:
            type: string
          description: From commit hash
        - name: to
          in: path
          required: true
          schema:
            type: string
          description: To commit hash
        - name: file_path
          in: path
          required: true
          schema:
            type: string
          description: Path to the file relative to the repository root
      responses:
        '200':
          description: File commit diff retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiffResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
          description: Document content (only included in certain endpoints)
        type:
          type: string
          enum: [document, folder, scrap]
        parent_id:
          type: string
          format: uuid
          nullable: true
        file_path:
          type: string
          nullable: true
          description: File system path for the document
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        visibility:
          type: string
          enum: [private, unlisted, public]
          description: Document visibility (private, unlisted, public)
        published_at:
          type: string
          format: date-time
          nullable: true
          description: Published date (only set when document is published)
        owner_username:
          type: string
          nullable: true
          description: Owner username (only included for published documents)
...
    UpdateDocumentRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        content:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
...
    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        file_size:
          type: integer
          format: int64
        mime_type:
          type: string
        document_id:
          type: string
          format: uuid
          nullable: true
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    FileResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            filename:
              type: string
            size:
              type: integer
            mime_type:
              type: string
            url:
              type: string

    # ===== Sharing =====
...
    GitCommit:
      type: object
      properties:
        id:
          type: string
          description: Commit hash
        message:
          type: string
          description: Commit message
        author_name:
          type: string
          description: Author name
        author_email:
          type: string
          description: Author email
        timestamp:
          type: string
          format: date-time
          description: Commit timestamp
        diff_stats:
          $ref: '#/components/schemas/DiffStats'
          nullable: true
    DiffStats:
      type: object
      properties:
        files_changed:
          type: integer
          format: int32
        insertions:
          type: integer
          format: int32
        deletions:
          type: integer
          format: int32
    DiffResult:
      type: object
      properties:
        file_path:
          type: string
          description: Path to the file being diffed
        diff_lines:
          type: array
          items:
            $ref: '#/components/schemas/DiffLine'
        old_content:
          type: string
          nullable: true
          description: Original file content (optional)
        new_content:
          type: string
          nullable: true
          description: New file content (optional)
    DiffLine:
      type: object
      properties:
        line_type:
          type: string
          enum: [added, deleted, context]
          description: Type of change for this line
        old_line_number:
          type: integer
          nullable: true
          description: Line number in the original file
        new_line_number:
          type: integer
          nullable: true
          description: Line number in the new file
        content:
          type: string
          description: The actual line content
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...